Move tests level up
