+++ .eslintrc
@@ -44,7 +44,7 @@
   },
   "overrides": [
     {
-      "files": ["*.spec.ts", "services/src/tests/**"],
+      "files": ["*.spec.ts", "services/tests/**"],
       "rules": {
         "@typescript-eslint/ban-types": "off",
         "@typescript-eslint/no-empty-function": "off",
