+++ services/jest.config.js
@@ -7,16 +7,16 @@ module.exports = {
       setupFiles: ['<rootDir>/setup-jest-unit.js'],
       testEnvironment: 'node',
       testMatch: ['<rootDir>/src/**/*.spec.ts'],
-      testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/', '<rootDir>/src/tests'],
+      testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/', '<rootDir>/tests'],
       testTimeout: 60000,
     },
     {
       displayName: 'Integration Tests',
       preset: 'ts-jest',
       rootDir: '.',
-      setupFiles: ['<rootDir>/src/tests/integration/setup.js'],
+      setupFiles: ['<rootDir>/tests/integration/setup.js'],
       testEnvironment: 'node',
-      testMatch: ['<rootDir>/src/tests/integration/**/*.spec.ts'],
+      testMatch: ['<rootDir>/tests/integration/**/*.spec.ts'],
       testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/dist/'],
       testTimeout: 60000,
     },
