+++ services/tests/e2e/docker-compose.yml
@@ -1,0 +1,48 @@
+version: '3.7'
+
+services:
+  gateway:
+    build:
+      context: ../..
+      dockerfile: deployment/docker/Dockerfile
+      args:
+        RUN_TESTS: 'false'
+    depends_on:
+      - minio
+    ports:
+      - '8080:80'
+    command: gateway
+    env_file: .env
+    environment:
+      PORT: 80
+
+  registry:
+    build:
+      context: ../..
+      dockerfile: deployment/docker/Dockerfile
+      args:
+        RUN_TESTS: 'false'
+    depends_on:
+      - minio
+    ports:
+      - '8090:80'
+    command: registry
+    env_file: .env
+    environment:
+      PORT: 80
+
+  minio:
+    image: minio/minio
+    ports:
+      - 9000:9000
+    entrypoint: sh
+    environment:
+      MINIO_ACCESS_KEY: ${AWS_ACCESS_KEY_ID}
+      MINIO_SECRET_KEY: ${AWS_SECRET_ACCESS_KEY}
+    command: >
+      -c "
+      mkdir -p /var/minio/${S3_RESOURCE_BUCKET_NAME} &&
+      echo '{\"schemas\":[],\"upstreams\": [],\"upstreamClientCredentials\": []}' > /var/minio/${S3_RESOURCE_BUCKET_NAME}/${S3_RESOURCE_OBJECT_KEY} &&
+      minio server /var/minio
+      "
+    env_file: .env
