+++ services/tests/integration/gateway/custom-scalars.spec.ts
@@ -3,7 +3,7 @@ import * as Rx from 'rxjs';
 import * as nock from 'nock';
 import { gql } from 'apollo-server-core';
 import { print } from 'graphql';
-import { createStitchGateway } from '../../../modules/gateway';
+import { createStitchGateway } from '../../../src/modules/gateway';
 import { beforeEachDispose } from '../before-each-dispose';
 
 const schema = {
